+++ uiobject.cpp
@@ -31,8 +31,8 @@ void UiObject::SetPosition(f32 x, f32 y) {
 }
 
 void UiObject::SetSize(u32 w, u32 h) {
-	size.x = w;
-	size.y = h;
+	size.x = (f32)w;
+	size.y = (f32)h;
 }
 
 void UiObject::SetVisible() {
